warn about external storages in guests_app.adoc
Guests have access to external storages, even when the files_external app is not in the whitelist.

I could not find anything to limit external storages to non-guest users. IMHO, that would be a sane default.

@pako81 Maybe this is a bug and should be fixed, rather than documented?

Not sure, if this warning should be here, or in the section about external storages instead.