clarify sharing bits
It is unclear, if only the listed numbers are supported.
I'd think, that 1+2+4+8+16 are meant as bit patters that can be combined as needed.
15, 31 are probably just examples.

What does 'default for public link shares' mean? Is the 1 bit always added for these? I don't think so, as this would prevent building an anonymous file drop link.